SwiftDocs-Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
redis==5.0.1
python-multipart==0.0.9
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
pdf2image==1.17.0
pytesseract==0.3.10
pymupdf==1.23.8
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80 